♻️ refactor/ZIP-61 : DTO 위치 및 변수명 수정 - 빌더 추가 - 코드리뷰 반영
