+++ src/main/java/com/zipte/platform/server/adapter/in/web/QuestionAPi.java
@@ -4,7 +4,7 @@ import com.zipte.platform.core.response.ApiResponse;
 import com.zipte.platform.core.response.pageable.PageRequest;
 import com.zipte.platform.core.response.pageable.PageResponse;
 import com.zipte.platform.server.adapter.in.web.dto.request.AnswerRequest;
-import com.zipte.platform.server.adapter.in.web.dto.request.QuestionAnswerDetailResponse;
+import com.zipte.platform.server.adapter.in.web.dto.response.QuestionAnswerDetailResponse;
 import com.zipte.platform.server.adapter.in.web.dto.request.QuestionRequest;
 import com.zipte.platform.server.adapter.in.web.dto.response.QuestionAnswerListResponse;
 import com.zipte.platform.server.application.in.community.AnswerUseCase;
