+++ src/main/java/com/zipte/platform/server/adapter/in/web/dto/request/QuestionRequest.java
@@ -1,9 +1,11 @@
 package com.zipte.platform.server.adapter.in.web.dto.request;
 
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
 import lombok.Data;
 
 @Data
+@Builder
 public class QuestionRequest {
 
     private Long userId;
