+++ src/main/java/com/zipte/platform/server/adapter/in/web/dto/response/QuestionAnswerDetailResponse.java
@@ -1,7 +1,5 @@
-package com.zipte.platform.server.adapter.in.web.dto.request;
+package com.zipte.platform.server.adapter.in.web.dto.response;
 
-import com.zipte.platform.server.adapter.in.web.dto.response.AnswerResponse;
-import com.zipte.platform.server.adapter.in.web.dto.response.QuestionResponse;
 import com.zipte.platform.server.domain.community.Answer;
 import com.zipte.platform.server.domain.community.Question;
 import lombok.Builder;
@@ -11,14 +9,14 @@ import java.util.List;
 @Builder
 public record QuestionAnswerDetailResponse(
         QuestionResponse question,
-        List<AnswerResponse> answer
+        List<AnswerResponse> answers
 ) {
 
     /// 정적 팩토리 메서드
     public static QuestionAnswerDetailResponse from(Question question, List<Answer> answers) {
         return QuestionAnswerDetailResponse.builder()
                 .question(QuestionResponse.from(question))
-                .answer(AnswerResponse.from(answers))
+                .answers(AnswerResponse.from(answers))
                 .build();
     }
