+++ src/main/java/com/zipte/platform/server/application/service/QuestionService.java
@@ -1,7 +1,7 @@
 package com.zipte.platform.server.application.service;
 
 import com.zipte.platform.core.response.ErrorCode;
-import com.zipte.platform.server.adapter.in.web.dto.request.QuestionAnswerDetailResponse;
+import com.zipte.platform.server.adapter.in.web.dto.response.QuestionAnswerDetailResponse;
 import com.zipte.platform.server.adapter.in.web.dto.request.QuestionRequest;
 import com.zipte.platform.server.adapter.in.web.dto.response.QuestionAnswerListResponse;
 import com.zipte.platform.server.application.in.community.QuestionUseCase;
